roles assigned

+++ lib/app/routers/users.py
@@ -7,7 +7,7 @@ router=APIRouter(prefix="/user",tags=["Users"])
 
 
 @router.get("/{id}")
-def get_user(id:int,db:Session=Depends(database.get_db)):
+def get_user(id:int,db:Session=Depends(database.get_db),role:str=Depends(oauth2.require_role(["admin"]))):
     user=db.query(models.Users).filter(models.Users.uid==id).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="User not found")
@@ -15,17 +15,16 @@ def get_user(id:int,db:Session=Depends(database.get_db)):
 
 
 
-@router.post("/create_user",response_model=schemas.UserOut)
-def create_user(user:schemas.CreateUser,db:Session=Depends(database.get_db)):
+@router.post("/create_user",response_model=schemas.UserOut,status_code=status.HTTP_201_CREATED)
+def create_user(user:schemas.CreateUser,db:Session=Depends(database.get_db),role:str=Depends(oauth2.require_role(["admin"]))):
     if(len(str(user.mobile))!=10):
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,detail="Invalid mobile number!")
     try:
         
         hashed_password=utils.hash_password(user.password)
         user.password=hashed_password
-        
+
         new_user=models.Users(**user.model_dump())
-        
         db.add(new_user)
         db.commit()
         db.refresh(new_user)
@@ -37,7 +36,7 @@ def create_user(user:schemas.CreateUser,db:Session=Depends(database.get_db)):
 
 
 @router.post("/delete_user/{id}")
-def delete_user(id:int,db:Session=Depends(database.get_db),current_user:int=Depends(oauth2.get_current_user)):
+def delete_user(id:int,db:Session=Depends(database.get_db),current_user:int=Depends(oauth2.get_current_user),role:str=Depends(oauth2.require_role(["admin","member"]))):
     
     user=db.query(models.Users).filter(models.Users.uid==id).first()
     if(not user):
@@ -45,6 +44,9 @@ def delete_user(id:int,db:Session=Depends(database.get_db),current_user:int=Depe
     
     if(user.uid!=current_user.uid):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail="Not authorized")
+    borrowed=db.query(models.IssueBook).filter(models.IssueBook.uid==current_user.uid).first()
+    if( borrowed):
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="You cannot delete your account while having borrowed books.")
     
     
     db.delete(user)
@@ -52,7 +54,7 @@ def delete_user(id:int,db:Session=Depends(database.get_db),current_user:int=Depe
     return {"message":"User deleted successfully"}
 
 @router.put("/update_user/{id}",status_code=status.HTTP_201_CREATED,response_model=schemas.UserOut)
-def update_user(id:int,user:schemas.CreateUser,db:Session=Depends(database.get_db),current_user:int=Depends(oauth2.get_current_user)):
+def update_user(id:int,user:schemas.CreateUser,db:Session=Depends(database.get_db),current_user:int=Depends(oauth2.get_current_user),role:str=Depends(oauth2.require_role(["admin","member"]))):
     
     if(len(str(user.mobile))!=10):
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,detail="Invalid mobile number!")
